+++ src/main/resources/db/migration/V24__create_invoice_number_counter_table.sql
@@ -1,6 +1,6 @@
 create table public.invoice_number_counter
 (
-    year                 smallint                         not null,
+    year                 int                              not null,
     counter              bigint                           not null,
     CONSTRAINT pk_invoice_number_counter_year PRIMARY KEY (year)
 );
